requirements.txt
pytest==8.3.4
pytest-playwright==1.49.1
pytest-dotenv==0.5.2
